3rd-party/lc3-google: fix warning for implicit ptrdiff_t to int conversions

+++ 3rd-party/lc3-google/src/bits.c
@@ -139,7 +139,7 @@ int lc3_check_bits(const struct lc3_bits *bits)
 static inline void accu_flush(
     struct lc3_bits_accu *accu, struct lc3_bits_buffer *buffer)
 {
-    int nbytes = LC3_MIN(accu->n >> 3,
+    int nbytes = (int) LC3_MIN(accu->n >> 3,
         LC3_MAX(buffer->p_bw - buffer->p_fw, 0));
 
     accu->n -= 8 * nbytes;
@@ -249,7 +249,7 @@ void lc3_flush_bits(struct lc3_bits *bits)
     struct lc3_bits_accu *accu = &bits->accu;
     struct lc3_bits_buffer *buffer = &bits->buffer;
 
-    int nleft = buffer->p_bw - buffer->p_fw;
+    int nleft = (int)(buffer->p_bw - buffer->p_fw);
     for (int n = 8 * nleft - accu->n; n > 0; n -= 32)
         lc3_put_bits(bits, 0, LC3_MIN(n, 32));
 
@@ -316,7 +316,7 @@ static inline int ac_get(struct lc3_bits_buffer *buffer)
 static inline void accu_load(struct lc3_bits_accu *accu,
     struct lc3_bits_buffer *buffer)
 {
-    int nbytes = LC3_MIN(accu->n >> 3, buffer->p_bw - buffer->start);
+    int nbytes = (int)LC3_MIN(accu->n >> 3, buffer->p_bw - buffer->start);
 
     accu->n -= 8 * nbytes;
